+++ mozaik_involvement_donation/models/__init__.py
@@ -1,2 +1,3 @@
+from . import partner_involvement
 from . import partner_involvement_category
 from . import res_partner
