+++ mozaik_membership/__manifest__.py
@@ -4,7 +4,7 @@
     "name": "Mozaik: Membership",
     "summary": """
         Manage memberships""",
-    "version": "14.0.1.1.2",
+    "version": "14.0.1.2.0",
     "license": "AGPL-3",
     "author": "ACSONE SA/NV",
     "website": "https://github.com/mozaik-association/mozaik",
